+++ src/buffer/view.rs
@@ -632,6 +632,18 @@ impl<T> BufferViewExt for BufferView<T> where T: Copy + Send + 'static {
         alloc.assert_unmapped(ctxt);
         alloc.get_id()
     }
+
+    fn bind_to(&self, ctxt: &mut CommandContext, ty: BufferType) {
+        let alloc = self.alloc.as_ref().unwrap();
+        alloc.assert_unmapped(ctxt);
+        alloc.bind(ctxt, ty);
+    }
+
+    fn indexed_bind_to(&self, ctxt: &mut CommandContext, ty: BufferType, index: gl::types::GLuint) {
+        let alloc = self.alloc.as_ref().unwrap();
+        alloc.assert_unmapped(ctxt);
+        alloc.indexed_bind(ctxt, ty, index, 0 .. alloc.get_size());
+    }
 }
 
 impl<'a, T> BufferViewSliceExt<'a> for BufferViewSlice<'a, T> where T: Copy + Send + 'static {
@@ -653,6 +665,17 @@ impl<'a, T> BufferViewExt for BufferViewSlice<'a, T> where T: Copy + Send + 'sta
         self.alloc.assert_unmapped(ctxt);
         self.alloc.get_id()
     }
+
+    fn bind_to(&self, ctxt: &mut CommandContext, ty: BufferType) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.bind(ctxt, ty);
+    }
+
+    fn indexed_bind_to(&self, ctxt: &mut CommandContext, ty: BufferType, index: gl::types::GLuint) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.indexed_bind(ctxt, ty, index, self.offset_bytes ..
+                                self.offset_bytes + self.num_elements * mem::size_of::<T>());
+    }
 }
 
 impl BufferViewExt for BufferViewAny {
@@ -663,6 +686,16 @@ impl BufferViewExt for BufferViewAny {
     fn get_buffer_id(&self, ctxt: &mut CommandContext) -> gl::types::GLuint {
         self.alloc.assert_unmapped(ctxt);
         self.alloc.get_id()
+    }
+
+    fn bind_to(&self, ctxt: &mut CommandContext, ty: BufferType) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.bind(ctxt, ty);
+    }
+
+    fn indexed_bind_to(&self, ctxt: &mut CommandContext, ty: BufferType, index: gl::types::GLuint) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.indexed_bind(ctxt, ty, index, 0 .. self.alloc.get_size());
     }
 }
 
@@ -684,6 +717,17 @@ impl<'a> BufferViewExt for BufferViewAnySlice<'a> {
     fn get_buffer_id(&self, ctxt: &mut CommandContext) -> gl::types::GLuint {
         self.alloc.assert_unmapped(ctxt);
         self.alloc.get_id()
+    }
+
+    fn bind_to(&self, ctxt: &mut CommandContext, ty: BufferType) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.bind(ctxt, ty);
+    }
+
+    fn indexed_bind_to(&self, ctxt: &mut CommandContext, ty: BufferType, index: gl::types::GLuint) {
+        self.alloc.assert_unmapped(ctxt);
+        self.alloc.indexed_bind(ctxt, ty, index, self.offset_bytes ..
+                                self.offset_bytes + self.elements_count * self.elements_size);
     }
 }