+++ src/context/extensions.rs
@@ -42,6 +42,8 @@ pub struct ExtensionsList {
     pub gl_arb_invalidate_subdata: bool,
     /// GL_ARB_map_buffer_range
     pub gl_arb_map_buffer_range: bool,
+    /// GL_ARB_multi_draw_indirect
+    pub gl_arb_multi_draw_indirect: bool,
     /// GL_ARB_occlusion_query
     pub gl_arb_occlusion_query: bool,
     /// GL_ARB_occlusion_query2
@@ -72,6 +74,8 @@ pub struct ExtensionsList {
     pub gl_arb_texture_storage: bool,
     /// GL_ARB_timer_query
     pub gl_arb_timer_query: bool,
+    /// GL_ARB_transform_feedback3
+    pub gl_arb_transform_feedback3: bool,
     /// GL_ARB_uniform_buffer_object
     pub gl_arb_uniform_buffer_object: bool,
     /// GL_ARB_vertex_array_object
@@ -80,6 +84,8 @@ pub struct ExtensionsList {
     pub gl_arb_vertex_buffer_object: bool,
     /// GL_ARB_vertex_shader
     pub gl_arb_vertex_shader: bool,
+    /// GL_ARM_rgba8
+    pub gl_arm_rgba8: bool,
     /// GL_ATI_meminfo
     pub gl_ati_meminfo: bool,
     /// GL_EXT_debug_marker
@@ -98,6 +104,8 @@ pub struct ExtensionsList {
     pub gl_ext_geometry_shader4: bool,
     /// GL_EXT_gpu_shader4
     pub gl_ext_gpu_shader4: bool,
+    /// GL_EXT_multi_draw_indirect
+    pub gl_ext_multi_draw_indirect: bool,
     /// GL_EXT_occlusion_query_boolean
     pub gl_ext_occlusion_query_boolean: bool,
     /// GL_EXT_packed_depth_stencil
@@ -128,6 +136,8 @@ pub struct ExtensionsList {
     pub gl_oes_depth_texture: bool,
     /// GL_OES_packed_depth_stencil
     pub gl_oes_packed_depth_stencil: bool,
+    /// GL_OES_rgb8_rgba8
+    pub gl_oes_rgb8_rgba8: bool,
     /// GL_OES_vertex_array_object
     pub gl_oes_vertex_array_object: bool,
 }
@@ -167,6 +177,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
         gl_arb_occlusion_query: false,
         gl_arb_occlusion_query2: false,
         gl_arb_map_buffer_range: false,
+        gl_arb_multi_draw_indirect: false,
         gl_arb_pixel_buffer_object: false,
         gl_arb_sampler_objects: false,
         gl_arb_shader_objects: false,
@@ -180,10 +191,12 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
         gl_arb_texture_rgb10_a2ui: false,
         gl_arb_texture_storage: false,
         gl_arb_timer_query: false,
+        gl_arb_transform_feedback3: false,
         gl_arb_uniform_buffer_object: false,
         gl_arb_vertex_array_object: false,
         gl_arb_vertex_buffer_object: false,
         gl_arb_vertex_shader: false,
+        gl_arm_rgba8: false,
         gl_ati_meminfo: false,
         gl_ext_debug_marker: false,
         gl_ext_direct_state_access: false,
@@ -193,6 +206,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
         gl_ext_framebuffer_srgb: false,
         gl_ext_geometry_shader4: false,
         gl_ext_gpu_shader4: false,
+        gl_ext_multi_draw_indirect: false,
         gl_ext_occlusion_query_boolean: false,
         gl_ext_packed_depth_stencil: false,
         gl_ext_texture_compression_s3tc: false,
@@ -208,6 +222,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
         gl_nvx_gpu_memory_info: false,
         gl_oes_depth_texture: false,
         gl_oes_packed_depth_stencil: false,
+        gl_oes_rgb8_rgba8: false,
         gl_oes_vertex_array_object: false,
     };
 
@@ -234,6 +249,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
             "GL_ARB_occlusion_query2" => extensions.gl_arb_occlusion_query2 = true,
             "GL_ARB_pixel_buffer_object" => extensions.gl_arb_pixel_buffer_object = true,
             "GL_ARB_map_buffer_range" => extensions.gl_arb_map_buffer_range = true,
+            "GL_ARB_multi_draw_indirect" => extensions.gl_arb_multi_draw_indirect = true,
             "GL_ARB_sampler_objects" => extensions.gl_arb_sampler_objects = true,
             "GL_ARB_shader_objects" => extensions.gl_arb_shader_objects = true,
             "GL_ARB_sync" => extensions.gl_arb_sync = true,
@@ -246,10 +262,12 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
             "GL_ARB_texture_rgb10_a2ui" => extensions.gl_arb_texture_rgb10_a2ui = true,
             "GL_ARB_texture_storage" => extensions.gl_arb_texture_storage = true,
             "GL_ARB_timer_query" => extensions.gl_arb_timer_query = true,
+            "GL_ARB_transform_feedback3" => extensions.gl_arb_transform_feedback3 = true,
             "GL_ARB_uniform_buffer_object" => extensions.gl_arb_uniform_buffer_object = true,
             "GL_ARB_vertex_array_object" => extensions.gl_arb_vertex_array_object = true,
             "GL_ARB_vertex_buffer_object" => extensions.gl_arb_vertex_buffer_object = true,
             "GL_ARB_vertex_shader" => extensions.gl_arb_vertex_shader = true,
+            "GL_ARM_rgba8" => extensions.gl_arm_rgba8 = true,
             "GL_ATI_meminfo" => extensions.gl_ati_meminfo = true,
             "GL_EXT_debug_marker" => extensions.gl_ext_debug_marker = true,
             "GL_EXT_direct_state_access" => extensions.gl_ext_direct_state_access = true,
@@ -259,6 +277,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
             "GL_EXT_framebuffer_sRGB" => extensions.gl_ext_framebuffer_srgb = true,
             "GL_EXT_geometry_shader4" => extensions.gl_ext_geometry_shader4 = true,
             "GL_EXT_gpu_shader4" => extensions.gl_ext_gpu_shader4 = true,
+            "GL_EXT_multi_draw_indirect" => extensions.gl_ext_multi_draw_indirect = true,
             "GL_EXT_occlusion_query_boolean" => extensions.gl_ext_occlusion_query_boolean = true,
             "GL_EXT_packed_depth_stencil" => extensions.gl_ext_packed_depth_stencil = true,
             "GL_EXT_texture_compression_s3tc" => extensions.gl_ext_texture_compression_s3tc = true,
@@ -274,6 +293,7 @@ pub unsafe fn get_extensions(gl: &gl::Gl, version: &Version) -> ExtensionsList {
             "GL_NVX_gpu_memory_info" => extensions.gl_nvx_gpu_memory_info = true,
             "GL_OES_depth_texture" => extensions.gl_oes_depth_texture = true,
             "GL_OES_packed_depth_stencil" => extensions.gl_oes_packed_depth_stencil = true,
+            "GL_OES_rgb8_rgba8" => extensions.gl_oes_rgb8_rgba8 = true,
             "GL_OES_vertex_array_object" => extensions.gl_oes_vertex_array_object = true,
             _ => ()
         }