+++ src/lib.rs
@@ -188,6 +188,13 @@ trait BufferViewExt {
 
     /// Makes sure that the buffer is available for operations and returns its ID.
     fn get_buffer_id(&self, &mut CommandContext) -> gl::types::GLuint;
+
+    /// Makes sure that the buffer is binded to a specific bind point.
+    fn bind_to(&self, &mut CommandContext, ty: buffer::BufferType);
+
+    /// Makes sure that the buffer is binded to a specific indexed bind point.
+    fn indexed_bind_to(&self, &mut CommandContext, ty: buffer::BufferType,
+                       index: gl::types::GLuint);
 }
 
 /// Internal trait for subbuffer slices.
@@ -245,6 +252,27 @@ trait QueryExt {
     /// Returns the type of this query object.
     fn get_type(&self) -> gl::types::GLenum;
 }
+
+/// Internal trait for textures.
+trait TextureExt {
+    /// Returns the bind point of the texture.
+    fn get_bind_point(&self) -> gl::types::GLenum;
+}
+
+/// Internal trait for transform feedback sessions.
+trait TransformFeedbackSessionExt {
+    /// Updates the state of OpenGL to make the transform feedback session current.
+    ///
+    /// The second parameter must be the primitive type of the input vertex data.
+    fn bind(&self, &mut CommandContext, index::PrimitiveType);
+
+    /// Ensures that transform feedback is disabled.
+    fn unbind(&mut CommandContext);
+
+    /// Ensures that a buffer isn't used by transform feedback.
+    fn ensure_buffer_out_of_transform_feedback(&mut CommandContext, gl::types::GLuint);
+}
+
 
 /// A raw value of a uniform. "Raw" means that it's passed directly with `glUniform`. Textures
 /// for example are just passed as integers.
@@ -607,7 +635,7 @@ pub trait Surface: Sized {
 /// Private trait for framebuffer-like objects that provide attachments.
 trait FboAttachments {
     /// Returns the list of attachments of this FBO, or `None` if it is the default framebuffer.
-    fn get_attachments(&self) -> Option<&fbo::FramebufferAttachments>;
+    fn get_attachments(&self) -> Option<&fbo::ValidatedAttachments>;
 }
 
 /// Error that can happen while drawing.
@@ -838,7 +866,7 @@ impl Surface for Frame {
 }
 
 impl FboAttachments for Frame {
-    fn get_attachments(&self) -> Option<&fbo::FramebufferAttachments> {
+    fn get_attachments(&self) -> Option<&fbo::ValidatedAttachments> {
         None
     }
 }