+++ src/texture/mod.rs
@@ -49,11 +49,11 @@ use image;
 use backend::Facade;
 
 use pixel_buffer::PixelBuffer;
-use uniforms::{UniformValue, AsUniformValue, Sampler};
+use uniforms::{UniformValue, AsUniformValue, UniformType, Sampler};
 use {Surface, GlObject};
 
 use FboAttachments;
-use fbo::FramebufferAttachments;
+use fbo::ValidatedAttachments;
 use Rect;
 use BlitTarget;
 use uniforms;
@@ -64,7 +64,7 @@ pub use image_format::{ClientFormat, TextureFormat};
 pub use image_format::{UncompressedFloatFormat, UncompressedIntFormat, UncompressedUintFormat};
 pub use image_format::{CompressedFormat, DepthFormat, DepthStencilFormat, StencilFormat};
 pub use image_format::{CompressedSrgbFormat, SrgbFormat};
-pub use self::any::TextureAny;
+pub use self::any::{TextureAny, TextureAnyMipmap, TextureType};
 pub use self::get_format::{InternalFormat, InternalFormatType};
 pub use self::pixel::PixelValue;
 
@@ -496,7 +496,7 @@ impl<'a> Surface for TextureSurface<'a> {
 }
 
 impl<'a> FboAttachments for TextureSurface<'a> {
-    fn get_attachments(&self) -> Option<&FramebufferAttachments> {
+    fn get_attachments(&self) -> Option<&ValidatedAttachments> {
         self.0.get_attachments()
     }
 }